+++ src/rcc.rs
@@ -43,13 +43,16 @@
 //!     // needed to configure specific peripherals.
 //!     // Looking at the peripheral function parameters
 //!     // should give more insight, which peripheral clock is needed.
-//!     .pclk1(16.mhz())
-//!     .pclk2(16.mhz())
+//!     .pclk1(12.mhz())
+//!     .pclk2(12.mhz())
 //!     // Freeze / apply the configuration and setup all clocks
-//!     .freeze(&mut flash.acr)
+//!     .freeze(&mut flash.acr);
 //! # }
-//!
 //! ```
+//!
+//! All fields can be omitted and will internally be set to a calculated default.
+//! For more details read the documentation of the [`CFGR`] methods to
+//! find out how to setup the clock.
 
 use crate::pac::{
     rcc::{self, cfgr, cfgr2},
@@ -196,15 +199,7 @@ mod usb_clocking {
     }
 }
 
-#[cfg(any(
-    feature = "stm32f302",
-    feature = "stm32f303",
-    feature = "stm32f373",
-    feature = "stm32f378",
-    feature = "stm32f328",
-    feature = "stm32f358",
-    feature = "stm32f398",
-))]
+#[cfg(not(any(feature = "stm32f301", feature = "stm32f318", feature = "stm32f334",)))]
 mod usb_clocking {
     use crate::pac::rcc::cfgr;
     use crate::rcc::PllConfig;
@@ -306,7 +301,7 @@ fn into_pll_mul(mul: u8) -> cfgr::PLLMUL_A {
         14 => cfgr::PLLMUL_A::MUL14,
         15 => cfgr::PLLMUL_A::MUL15,
         16 => cfgr::PLLMUL_A::MUL16,
-        _ => unreachable!(),
+        _ => crate::unreachable!(),
     }
 }
 
@@ -329,7 +324,7 @@ fn into_pre_div(div: u8) -> cfgr2::PREDIV_A {
         14 => cfgr2::PREDIV_A::DIV14,
         15 => cfgr2::PREDIV_A::DIV15,
         16 => cfgr2::PREDIV_A::DIV16,
-        _ => unreachable!(),
+        _ => crate::unreachable!(),
     }
 }
 
@@ -373,6 +368,11 @@ impl CFGR {
     }
 
     /// Sets a frequency for the `APB1` bus
+    ///
+    /// - Maximal supported frequency: 36 Mhz
+    ///
+    /// If not manually set, it will be set to [`CFGR::sysclk`] frequency
+    /// or [`CFGR::sysclk`] frequency / 2, if [`CFGR::sysclk`] > 36 Mhz
     pub fn pclk1<F>(mut self, freq: F) -> Self
     where
         F: Into<Hertz>,
@@ -382,6 +382,16 @@ impl CFGR {
     }
 
     /// Sets a frequency for the `APB2` bus
+    ///
+    /// # Resolution and Limits
+    ///
+    /// Following is true for devices **except**, as these allow finer resolutions
+    /// even when using the internal oscillator:
+    ///
+    ///     [stm32f302xd,stm32f302xe,stm32f303xd,stm32f303xe,stm32f398]
+    ///
+    /// - Maximal supported frequency with HSE: 72 Mhz
+    /// - Maximal supported frequency without HSE: 64 Mhz
     pub fn pclk2<F>(mut self, freq: F) -> Self
     where
         F: Into<Hertz>,
@@ -391,6 +401,19 @@ impl CFGR {
     }
 
     /// Sets the system (core) frequency
+    ///
+    /// # Resolution and Limits
+    ///
+    /// Following is true for devices **except**, as these allow finer resolutions
+    /// even when using the internal oscillator:
+    ///
+    ///     [stm32f302xd,stm32f302xe,stm32f303xd,stm32f303xe,stm32f398]
+    ///
+    /// - Maximal supported frequency with `HSE`: 72 Mhz
+    /// - Maximal supported frequency without `HSE`: 64 Mhz
+    ///
+    /// If [`CFGR::hse`] is not used, therefor `HSI / 2` is used.
+    /// Only multiples of (HSI / 2) (4 Mhz) are allowed.
     pub fn sysclk<F>(mut self, freq: F) -> Self
     where
         F: Into<Hertz>,
@@ -443,22 +466,22 @@ impl CFGR {
             }
 
             // PLL_MUL maximal value is 16
-            assert!(multiplier <= 16);
+            crate::assert!(multiplier <= 16);
 
             // PRE_DIV maximal value is 16
-            assert!(divisor <= 16);
+            crate::assert!(divisor <= 16);
 
             (multiplier, Some(divisor))
         }
         // HSI division is always divided by 2 and has no adjustable division
         else {
             let pll_mul = sysclk / pllsrcclk;
-            assert!(pll_mul <= 16);
+            crate::assert!(pll_mul <= 16);
             (pll_mul, None)
         };
 
         let sysclk = (pllsrcclk / pll_div.unwrap_or(1)) * pll_mul;
-        assert!(sysclk <= 72_000_000);
+        crate::assert!(sysclk <= 72_000_000);
 
         let pll_src = if self.hse.is_some() {
             cfgr::PLLSRC_A::HSE_DIV_PREDIV
@@ -518,16 +541,16 @@ impl CFGR {
             }
 
             // PLL_MUL maximal value is 16
-            assert!(multiplier <= 16);
+            crate::assert!(multiplier <= 16);
 
             // PRE_DIV maximal value is 16
-            assert!(divisor <= 16);
+            crate::assert!(divisor <= 16);
 
             (multiplier, divisor)
         };
 
         let sysclk = (pllsrcclk / pll_div) * pll_mul;
-        assert!(sysclk <= 72_000_000);
+        crate::assert!(sysclk <= 72_000_000);
 
         // Select hardware clock source of the PLL
         // TODO Check whether HSI_DIV2 could be useful
@@ -598,7 +621,7 @@ impl CFGR {
         let (hpre_bits, hpre) =
             self.hclk
                 .map_or((cfgr::HPRE_A::DIV1, 1), |hclk| match sysclk / hclk {
-                    0 => unreachable!(),
+                    0 => crate::unreachable!(),
                     1 => (cfgr::HPRE_A::DIV1, 1),
                     2 => (cfgr::HPRE_A::DIV2, 2),
                     3..=5 => (cfgr::HPRE_A::DIV4, 4),
@@ -612,12 +635,12 @@ impl CFGR {
 
         let hclk: u32 = sysclk / hpre;
 
-        assert!(hclk <= 72_000_000);
+        crate::assert!(hclk <= 72_000_000);
 
-        let (ppre1_bits, ppre1) =
+        let (mut ppre1_bits, mut ppre1) =
             self.pclk1
                 .map_or((cfgr::PPRE1_A::DIV1, 1), |pclk1| match hclk / pclk1 {
-                    0 => unreachable!(),
+                    0 => crate::unreachable!(),
                     1 => (cfgr::PPRE1_A::DIV1, 1),
                     2 => (cfgr::PPRE1_A::DIV2, 2),
                     3..=5 => (cfgr::PPRE1_A::DIV4, 4),
@@ -625,14 +648,24 @@ impl CFGR {
                     _ => (cfgr::PPRE1_A::DIV16, 16),
                 });
 
-        let pclk1 = hclk / u32::from(ppre1);
+        let mut pclk1 = hclk / u32::from(ppre1);
 
-        assert!(pclk1 <= 36_000_000);
+        // This ensures, that no panic happens, when
+        // pclk1 is not manually set.
+        // As hclk highest value is 72.mhz()
+        // dividing by 2 should always be sufficient
+        if self.pclk1.is_none() && pclk1 > 36_000_000 {
+            ppre1_bits = cfgr::PPRE1_A::DIV2;
+            ppre1 = 2;
+            pclk1 = hclk / u32::from(ppre1);
+        }
+
+        crate::assert!(pclk1 <= 36_000_000);
 
         let (ppre2_bits, ppre2) =
             self.pclk2
                 .map_or((cfgr::PPRE2_A::DIV1, 1), |pclk2| match hclk / pclk2 {
-                    0 => unreachable!(),
+                    0 => crate::unreachable!(),
                     1 => (cfgr::PPRE2_A::DIV1, 1),
                     2 => (cfgr::PPRE2_A::DIV2, 2),
                     3..=5 => (cfgr::PPRE2_A::DIV4, 4),
@@ -642,7 +675,7 @@ impl CFGR {
 
         let pclk2 = hclk / u32::from(ppre2);
 
-        assert!(pclk2 <= 72_000_000);
+        crate::assert!(pclk2 <= 72_000_000);
 
         // Adjust flash wait states according to the
         // HCLK frequency (cpu core clock)
@@ -762,6 +795,14 @@ impl Clocks {
     }
 
     /// Returns whether the USBCLK clock frequency is valid for the USB peripheral
+    ///
+    /// If the micrcontroller does support USB, 48 Mhz or 72 Mhz have to be used
+    /// and the [`CFGR::hse`] must be used.
+    ///
+    ///  The APB1 / [`CFGR::pclk1`] clock must have a minimum frequency of 10 MHz to avoid data
+    ///  overrun/underrun problems. [RM0316 32.5.2][RM0316]
+    ///
+    /// [RM0316]: https://www.st.com/resource/en/reference_manual/dm00043574.pdf
     pub fn usbclk_valid(&self) -> bool {
         self.usbclk_valid
     }